requirements.txt
python-pptx==0.6.21
groq==0.4.2
streamlit==1.32.0 